+++ src/server/services/about/controllers/index.js
@@ -1,0 +1,9 @@
+import { serviceController } from '~/src/server/services/about/controllers/service'
+import { serviceListController } from '~/src/server/services/about/controllers/service-list'
+import { serviceCreateStatusController } from '~/src/server/services/about/controllers/service-create-status'
+
+export {
+  serviceController,
+  serviceListController,
+  serviceCreateStatusController
+}
