+++ src/server/services/about/views/service-status.njk
@@ -20,7 +20,7 @@
     {% endif %}
 
       {% block xhrContent %}
-        {% include "services/partials/service-status.njk" %}
+        {% include "services/about/partials/service-status.njk" %}
       {% endblock %}
 
     {% if not isSuccess %}
