--- src/server/services/controllers/index.js
@@ -1,21 +1,0 @@
-import { serviceController } from '~/src/server/services/controllers/service'
-import { serviceListController } from '~/src/server/services/controllers/service-list'
-import { serviceCreateStatusController } from '~/src/server/services/controllers/service-create-status'
-import {
-  allSecretsController,
-  environmentSecretsController,
-  updateSecretController,
-  createSecretController,
-  updateSecretFormController
-} from '~/src/server/services/controllers/secrets'
-
-export {
-  serviceController,
-  serviceListController,
-  serviceCreateStatusController,
-  allSecretsController,
-  environmentSecretsController,
-  updateSecretController,
-  createSecretController,
-  updateSecretFormController
-}
