--- src/server/services/controllers/secrets/index.js
@@ -1,13 +1,0 @@
-import { allSecretsController } from '~/src/server/services/controllers/secrets/all'
-import { environmentSecretsController } from '~/src/server/services/controllers/secrets/environment'
-import { updateSecretController } from '~/src/server/services/controllers/secrets/update'
-import { createSecretController } from '~/src/server/services/controllers/secrets/create'
-import { updateSecretFormController } from '~/src/server/services/controllers/secrets/update-form'
-
-export {
-  allSecretsController,
-  environmentSecretsController,
-  updateSecretController,
-  createSecretController,
-  updateSecretFormController
-}
