--- src/server/services/helpers/fetch/fetch-repositories.js
@@ -1,13 +1,0 @@
-import { config } from '~/src/config'
-import { fetcher } from '~/src/server/common/helpers/fetch/fetcher'
-
-async function fetchRepositories(teamId = null) {
-  const endpoint =
-    config.get('portalBackendApiUrl') +
-    `/repositories${teamId ? `?team=${teamId}` : ''}`
-
-  const { json } = await fetcher(endpoint)
-  return json
-}
-
-export { fetchRepositories }
