+++ src/server/services/secrets/controllers/environment.js
@@ -1,48 +1,50 @@
-import Joi from 'joi'
 import Boom from '@hapi/boom'
 import { kebabCase, upperFirst } from 'lodash'
 
-import { config } from '~/src/config'
 import { fetchSecrets } from '~/src/server/common/helpers/fetch/fetch-secrets'
 import { provideService } from '~/src/server/services/helpers/pre/provide-service'
-import { secretParamsValidation } from '~/src/server/services/helpers/schema/secret-validation'
+import { environmentSecrets } from '~/src/server/services/secrets/transformers/environment-secrets'
+import { secretParamsValidation } from '~/src/server/services/secrets/helpers/schema/secret-validation'
 
-const immutableKeys = config.get('platformGlobalSecretKeys')
-
-const updateSecretFormController = {
+const environmentSecretsController = {
   options: {
-    id: 'services/{serviceId}/secrets/{environment}/update',
+    id: 'services/{serviceId}/secrets/{environment}',
     pre: [provideService],
     validate: {
       params: secretParamsValidation,
-      query: Joi.object({
-        secretKey: Joi.string()
-          .not(...immutableKeys)
-          .min(1)
-          .max(256)
-          .required()
-      }),
       failAction: () => Boom.boomify(Boom.notFound())
     }
   },
   handler: async (request, h) => {
+    const environment = request.params.environment
     const service = request.pre.service
+    const serviceName = service.serviceName
     const team = service?.teams?.at(0)
     const teamId = team?.teamId
-    const serviceName = service.serviceName
-    const environment = request.params?.environment
-    const secretKey = request.query?.secretKey
     const formattedEnvironment = upperFirst(kebabCase(environment))
     const secrets = await fetchSecrets(environment, serviceName)
 
-    return h.view('services/views/secrets/update-form', {
-      pageTitle: `${serviceName} - Update secret`,
+    const {
+      serviceSecrets,
+      platformSecrets,
+      shouldPoll,
+      successMessage,
+      exceptionMessage,
+      isSecretsSetup
+    } = environmentSecrets(secrets)
+
+    return h.view('services/secrets/views/environment', {
+      pageTitle: `${serviceName} - Secrets - ${formattedEnvironment}`,
       heading: serviceName,
       service,
       teamId,
       environment,
-      secretKey,
-      isSecretsSetup: secrets !== null,
+      platformSecrets,
+      serviceSecrets,
+      shouldPoll,
+      successMessage,
+      exceptionMessage,
+      isSecretsSetup,
       breadcrumbs: [
         {
           text: 'Services',
@@ -57,15 +59,11 @@ const updateSecretFormController = {
           href: `/services/${serviceName}/secrets`
         },
         {
-          text: formattedEnvironment,
-          href: `/services/${serviceName}/secrets/${environment}`
-        },
-        {
-          text: `Update secret`
+          text: formattedEnvironment
         }
       ]
     })
   }
 }
 
-export { updateSecretFormController }
+export { environmentSecretsController }
