+++ src/server/services/secrets/controllers/index.js
@@ -1,0 +1,13 @@
+import { allSecretsController } from '~/src/server/services/secrets/controllers/all'
+import { environmentSecretsController } from '~/src/server/services/secrets/controllers/environment'
+import { updateSecretController } from '~/src/server/services/secrets/controllers/update'
+import { createSecretController } from '~/src/server/services/secrets/controllers/create'
+import { updateSecretFormController } from '~/src/server/services/secrets/controllers/update-form'
+
+export {
+  allSecretsController,
+  environmentSecretsController,
+  updateSecretController,
+  createSecretController,
+  updateSecretFormController
+}
