+++ src/server/services/secrets/helpers/add-service-owner-scope.js
