+++ src/server/services/secrets/views/environment.njk
@@ -194,7 +194,7 @@
 
       <section class="app-grid-column-one-half-from-desktop-big">
         {% if isSecretsSetup %}
-          {% include "services/partials/secrets/create-form.njk" %}
+          {% include "services/secrets/partials/create-form.njk" %}
         {% endif %}
       </section>
     </div>
